EY-5073 lagre inntekter fra skatt

+++ apps/etterlatte-behandling/src/main/kotlin/behandling/etteroppgjoer/sigrun/SigrunKlient.kt
@@ -79,7 +79,7 @@ data class PensjonsgivendeInntektResponse(
 
 fun PensjonsgivendeInntektAarResponse.fromResponse() =
     PensjonsgivendeInntektFraSkatt(
-        inntektsaar = inntektsaar,
+        inntektsaar = inntektsaar.toInt(),
         inntekter =
             pensjonsgivendeInntekt.map {
                 PensjonsgivendeInntekt(
